+++ packages/adapters-library/src/adapters/convex/products/pool/convexPoolAdapter.ts
@@ -1,8 +1,6 @@
 import { getAddress } from 'ethers'
-import {
-  LpStakingAdapter,
-  LpStakingProtocolMetadata,
-} from '../../../../core/adapters/LpStakingProtocolAdapter'
+
+import { SimplePoolAdapter } from '../../../../core/adapters/SimplePoolAdapter'
 import { Chain } from '../../../../core/constants/chains'
 import {
   CacheToFile,
@@ -21,21 +19,76 @@ import {
   PositionType,
   ProtocolDetails,
   ProtocolPosition,
+  TokenBalance,
+  TokenType,
+  Underlying,
   UnderlyingReward,
+  UnwrappedTokenExchangeRate,
 } from '../../../../types/adapter'
+import { Erc20Metadata } from '../../../../types/erc20Metadata'
 import { CONVEX_FACTORY_ADDRESS } from '../../common/constants'
 import { ConvexFactory__factory } from '../../contracts'
 
-/**
- * First version of Convex had additional token which needed to be staked to accrue rewards
- */
-export class ConvexPoolAdapter extends LpStakingAdapter {
+const PRICE_PEGGED_TO_ONE = 1
+
+export type ExtraRewardToken = {
+  token: Erc20Metadata
+  manager: string
+}
+
+export type AdditionalMetadata = {
+  underlyingTokens: Erc20Metadata[]
+  extraRewardTokens: ExtraRewardToken[]
+}
+
+export class ConvexPoolAdapter extends SimplePoolAdapter<AdditionalMetadata> {
   productId = 'pool'
 
   adapterSettings = {
     enablePositionDetectionByProtocolTokenTransfer: false,
     includeInUnwrap: true,
     version: 2,
+  }
+
+  protected async unwrapProtocolToken(
+    protocolTokenMetadata: Erc20Metadata,
+    _blockNumber?: number,
+  ): Promise<UnwrappedTokenExchangeRate[]> {
+    const underlyingTokens = await this.fetchUnderlyingTokensMetadata(
+      protocolTokenMetadata.address,
+    )
+
+    const pricePerShareRaw = BigInt(
+      PRICE_PEGGED_TO_ONE * 10 ** protocolTokenMetadata.decimals,
+    )
+
+    return [
+      {
+        ...underlyingTokens[0]!,
+        type: TokenType.Underlying,
+        underlyingRateRaw: pricePerShareRaw,
+      },
+    ]
+  }
+
+  protected async getUnderlyingTokenBalances({
+    protocolTokenBalance,
+  }: {
+    userAddress: string
+    protocolTokenBalance: TokenBalance
+    blockNumber?: number
+  }): Promise<Underlying[]> {
+    const underlyingTokens = await this.fetchUnderlyingTokensMetadata(
+      protocolTokenBalance.address,
+    )
+
+    const underlyingTokenBalance = {
+      ...underlyingTokens[0]!,
+      balanceRaw: protocolTokenBalance.balanceRaw,
+      type: TokenType.Underlying,
+    }
+
+    return [underlyingTokenBalance]
   }
 
   async getRewardPositionsLpStakingAdapter(
@@ -61,9 +114,7 @@ export class ConvexPoolAdapter extends LpStakingAdapter {
   }
 
   @CacheToFile({ fileKey: 'metadata' })
-  async getProtocolTokens(): Promise<
-    ProtocolToken<LpStakingProtocolMetadata>[]
-  > {
+  async getProtocolTokens(): Promise<ProtocolToken<AdditionalMetadata>[]> {
     const convexFactory = ConvexFactory__factory.connect(
       CONVEX_FACTORY_ADDRESS,
       this.provider,
@@ -71,7 +122,7 @@ export class ConvexPoolAdapter extends LpStakingAdapter {
 
     const pools = await convexFactory.poolLength()
 
-    const metadata: ProtocolToken<LpStakingProtocolMetadata>[] = []
+    const metadata: ProtocolToken<AdditionalMetadata>[] = []
     await Promise.all(
       Array.from({ length: Number(pools) }, async (_, i) => {
         const convexData = await convexFactory.poolInfo(i)
